Move sync_badge into its own folder
