+++ x-pack/solutions/observability/plugins/apm/public/components/shared/trace_waterfall/badges/sync_badge/helper.ts
