+++ x-pack/solutions/observability/plugins/apm/public/components/shared/trace_waterfall/badges/sync_badge/sync_badge.stories.tsx
@@ -7,8 +7,8 @@
 
 import type { Meta, StoryObj } from '@storybook/react';
 import React from 'react';
-import { SyncBadge } from './sync_badge';
-import { agentsSyncMap } from './constants';
+import { SyncBadge } from '.';
+import { agentsSyncMap } from './helper';
 
 const columnStyle: React.CSSProperties = {
   display: 'flex',
